Get rid of LargeByteArray.

+++ src/main/java/org/pcj/internal/network/LoopbackMessageBytesStream.java
@@ -145,7 +145,7 @@ public class LoopbackMessageBytesStream implements AutoCloseable {
                 } else if (byteBuffer.hasRemaining() == false) {
                     queue.poll();
                 } else {
-                    return ((int) byteBuffer.get()) & 0xFF;
+                    return byteBuffer.get() & 0xFF;
                 }
             }
         }
